Added messaging config

+++ packages/optimize/android/src/main/java/com/adobe/marketing/mobile/reactnative/optimize/RCTAEPOptimizeUtil.java
@@ -72,7 +72,7 @@ class RCTAEPOptimizeUtil {
         if (offer.getMeta() != null) {
             offerWritableMap.putMap("meta", convertMapToWritableMap(new HashMap<String, Object>(offer.getMeta())));
         }
-        offerWritableMap.putInt("score", offer.getScore());
+        offerWritableMap.putInt("score", (int) offer.getScore());
 
         final WritableMap dataWritableMap = new WritableNativeMap();
         dataWritableMap.putString("id", offer.getId());
